+++ serde_with/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.18.0] - 2026-03-13
+
+### Added
+
+* Support `OneOrMany` with more sequence and set types (#929)
+
 ### Changed
 
 * Bump MSRV to 1.88 due to the `darling` dependency